Decouple bootstrap and sync
